# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "ActualScoreUpdateRequest": {
        "properties": {
          "actual": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Actual",
            "type": "number"
          },
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic",
          "actual"
        ],
        "title": "ActualScoreUpdateRequest",
        "type": "object"
      },
      "ActualScoreUpdateResponse": {
        "properties": {
          "actual": {
            "title": "Actual",
            "type": "number"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "updated": {
            "title": "Updated",
            "type": "boolean"
          }
        },
        "required": [
          "course_id",
          "topic",
          "actual",
          "updated"
        ],
        "title": "ActualScoreUpdateResponse",
        "type": "object"
      },
...
      "CalibrationRatingListResponse": {
        "properties": {
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "ratings": {
            "items": {
              "$ref": "#/components/schemas/CalibrationRatingResponse"
            },
            "title": "Ratings",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "ratings"
        ],
        "title": "CalibrationRatingListResponse",
        "type": "object"
      },
      "CalibrationRatingRequest": {
        "properties": {
          "course_id": {
            "exclusiveMinimum": 0.0,
            "title": "Course Id",
            "type": "integer"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "topic": {
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "course_id",
          "topic",
          "rating"
        ],
        "title": "CalibrationRatingRequest",
        "type": "object"
      },
      "CalibrationRatingResponse": {
        "properties": {
          "actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "calibration_error": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calibration Error"
          },
          "course_id": {
            "title": "Course Id",
            "type": "integer"
          },
          "difficulty_level": {
            "title": "Difficulty Level",
            "type": "string"
          },
          "is_blind_spot": {
            "title": "Is Blind Spot",
            "type": "boolean"
          },
          "predicted": {
            "title": "Predicted",
            "type": "number"
          },
          "rated_at": {
            "title": "Rated At",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "course_id",
          "predicted",
          "actual",
          "rated_at",
          "calibration_error",
          "is_blind_spot",
          "difficulty_level"
        ],
        "title": "CalibrationRatingResponse",
        "type": "object"
      },
      "CalibrationRatingSavedResponse": {
        "properties": {
          "rating": {
            "$ref": "#/components/schemas/CalibrationRatingResponse"
          }
        },
        "required": [
          "rating"
        ],
        "title": "CalibrationRatingSavedResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonPrimitive"
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonPrimitive": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sophia API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/calibration/actual-score": {
      "patch": {
        "operationId": "updateCalibrationActualScore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualScoreUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActualScoreUpdateResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Patch Actual Score",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/calibration/blind-spots": {
      "get": {
        "operationId": "listCalibrationBlindSpots",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationRatingListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Calibration Blind Spots",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/calibration/ratings": {
      "get": {
        "operationId": "listCalibrationRatings",
        "parameters": [
          {
            "in": "query",
            "name": "course_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Course Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "topic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationRatingListResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List Calibration Ratings",
        "tags": [
          "calibration"
        ]
      },
      "post": {
        "operationId": "saveCalibrationConfidenceRating",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationRatingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationRatingSavedResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content",
            "headers": {
              "X-Request-ID": {
                "description": "Request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Save Calibration Rating",
        "tags": [
          "calibration"
        ]
      }
    },
...
  }
}